move relevant types to route-patterns, add clarity for group keys
